+++ src/model/PremiumUser.java
@@ -63,7 +63,7 @@ public class PremiumUser extends User implements Payable {
     }
 
     @Override
-    public String generatePayment(int value) {
+    public String generatePayment(double value) {
         var info = "A receipt by $" + value + " will be sent to your email";
         int month = Calendar.getInstance().get(Calendar.MONTH);
         payments[month] = value;
